splitted merkle tree and conventional tree into modules
